requirements.txt
asgiref==3.6.0
certifi==2023.5.7
Django==3.2.6
django-cors-headers==4.0.0
django-elasticsearch-dsl==7.2.0
django-environ==0.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
